Add files via upload
in this assignment our objective was to create a random dataset and teach a perceptron how to sort it into its correct categories. A perceptron is like a mini-brain cell, or neuron, that helps a computer learn and make decisions based on data it receives. In this case the data is how many points it correctly classified or misclassified.